+++ docs/changes.rst
@@ -1,6 +1,15 @@
 Geofront Changelog
 ==================
 
+Version 0.1.1
+-------------
+
+To be released.
+
+- Fixed :exc:`TypeError` that rises when :class:`CloudMasterKeyStore
+  <geofront.masterkey.CloudMasterKeyStore>` is used with AWS S3 driver.
+
+
 Version 0.1.0
 -------------